+++ source/app/service-providers/commands/language-tool.ts
@@ -18,6 +18,7 @@ import { URLSearchParams } from 'url'
 import { app } from 'electron'
 import { trans } from '@common/i18n-main'
 import type { AppServiceContainer } from 'source/app/app-service-container'
+import { type LanguageToolIgnoredRuleEntry } from '../config/get-config-template'
 
 export interface Annotation {
   text?: string,
@@ -29,32 +30,130 @@ export interface AnnotationData {
   annotation: Annotation[]
 }
 
+/**
+ * This interface describes a single match reported by the LanguageTool API.
+ */
 export interface LanguageToolAPIMatch {
-  message: string // Long message (English)
+  /**
+   * The long message (usually localized).
+   */
+  message: string
+  /**
+   * A short version of this message (usually localized). NOTE: May be an empty
+   * string.
+   */
   shortMessage: string
+  /**
+   * A list of possible replacements to fix this match.
+   */
   replacements: Array<{ value: string }>
+  /**
+   * The offset in the character stream for this match.
+   */
   offset: number
+  /**
+   * The length of this match (end point is offset + length).
+   */
   length: number
+  /**
+   * Contextual information for this match.
+   */
   context: {
-    text: string // Abbreviated text
+    /**
+     * An abridged version of the text surrounding this match.
+     */
+    text: string
+    /**
+     * The start offset of the context.
+     */
     offset: number
+    /**
+     * The length of the context.
+     */
     length: number
   }
-  sentence: string // Full text
+  /**
+   * The full sentence surrounding the match.
+   */
+  sentence: string
+  /**
+   * A bit of information on the type of match.
+   */
   type: {
+    /**
+     * The typeName is something different from the issueType. I am not entirely
+     * sure what it describes.
+     */
     typeName: string
   }
+  /**
+   * General information about the rule that produced this match.
+   */
   rule: {
+    /**
+     * A unique ID of this match, this can be used to ignore this rule for the
+     * coming calls to the API.
+     */
     id: string
+    /**
+     * This is a (usually localized) description of the rule.
+     */
     description: string
+    /**
+     * This is the issue type of this rule (not the same as typeName).
+     */
     issueType: string // misspelling, grammar, etc
+    /**
+     * The broader category of this rule.
+     */
     category: {
+      /**
+       * The ID of the category
+       */
       id: string
+      /**
+       * The (localized) name of the category.
+       */
       name: string
     }
+    /**
+     * Whether the user is using premium.
+     */
     isPremium: boolean
+    /**
+     * The confidence of LT that this match is a violation of this rule. Appears
+     * to be bound between 0 and 1. Might be useful to filter out messages even
+     * more based on how confident the API is.
+     */
+    confidence: number
+    /**
+     * An optional filename for where the rule is coming from. This does not
+     * exist on all matches, which makes me suspect that this is only present
+     * for rules defined in the `grammar.xml` file of the corresponding
+     * language, but not for rules defined directly in the Java source.
+     */
+    sourceFile?: string
+    /**
+     * Appears to belong to the `sourceFile` argument; maybe it describes a
+     * specific ID within the `grammar.xml` file.
+     */
+    subID?: string
+    /**
+     * A list of tags for this rule. I've seen it containing the string "picky",
+     * which indicates that this can be used to identify whether a match has
+     * been produced by the picky setting of the API.
+     */
+    tags?: string[]
   }
+  /**
+   * Appears to indicate if we should ignore this match in case we're dealing
+   * with an incomplete sentence.
+   */
   ignoreForIncompleteSentence: boolean
+  /**
+   * I am not entirely certain what this parameter does, but it appears to be 0
+   * or -1.
+   */
   contextForSureMatch: number
 }
 
@@ -111,9 +210,12 @@ async function fetchSupportedLanguages (server: string): Promise<string[]> {
   return allLanguages
 }
 
+export type LanguageToolLinterRequest = { data: AnnotationData, language: string }
+export type LanguageToolLinterResponse = [LanguageToolAPIResponse, supportedLanguages: string[]]|string|undefined
+
 export default class LanguageTool extends ZettlrCommand {
   constructor (app: AppServiceContainer) {
-    super(app, 'run-language-tool')
+    super(app, [ 'run-language-tool', 'add-language-tool-ignore-rule' ])
   }
 
   /**
@@ -124,12 +226,21 @@ export default class LanguageTool extends ZettlrCommand {
    * to-be-expected happened (such as: the service is not even turned on in the
    * settings).
    *
-   * @param   {string}                       evt  The event
-   * @param   {string}                       arg  The text to check
+   * @param   {'add-language-tool-ignore-rule'|'run-language-tool'}     evt  The event
+   * @param   {LanguageToolLinterRequest|LanguageToolIgnoredRuleEntry}  arg  The text to check
    *
-   * @return  {Promise<LanguageToolAPIResponse|string|undefined>}       The result
+   * @return  {Promise<LanguageToolLinterResponse|boolean>}                  The result
    */
-  async run (evt: string, arg: { data: AnnotationData, language: string, disabledRules: string[] }): Promise<[LanguageToolAPIResponse, string[]]|string|undefined> {
+  async run (evt: 'add-language-tool-ignore-rule', arg: LanguageToolIgnoredRuleEntry): Promise<boolean>
+  async run (evt: 'run-language-tool', arg: LanguageToolLinterRequest): Promise<LanguageToolLinterResponse>
+  async run (evt: 'add-language-tool-ignore-rule'|'run-language-tool', arg: LanguageToolLinterRequest|LanguageToolIgnoredRuleEntry): Promise<LanguageToolLinterResponse|boolean> {
+    // Check if we only should add a rule to our ignore list
+    if (evt === 'add-language-tool-ignore-rule' && 'id' in arg) {
+      return this.addIgnoreRule(arg)
+    } else if (!('data' in arg)) {
+      throw new Error('Wrong call signature for language tool command. Either you called add-ignore-rule with a LanguageTool lint request, or run-language-tool with an ignore rule.')
+    }
+
     const {
       active,
       level,
@@ -138,15 +249,17 @@ export default class LanguageTool extends ZettlrCommand {
       username,
       apiKey,
       customServer,
-      provider
+      provider,
+      ignoredRules
     } = this._app.config.getConfig().editor.lint.languageTool
-    const { data, language, disabledRules } = arg
+    const { data, language } = arg
 
     if (!active) {
       return undefined // LanguageTool is not active
     }
 
-    const searchParams = new URLSearchParams({ language, data: JSON.stringify(data), level, disabledRules: disabledRules.join(',') })
+    const disabledRules = ignoredRules.map(r => r.id).join(',')
+    const searchParams = new URLSearchParams({ language, data: JSON.stringify(data), level, disabledRules })
 
     // If the user has set the mother tongue, add it to the params
     if (motherTongue.trim() !== '') {
@@ -214,5 +327,18 @@ export default class LanguageTool extends ZettlrCommand {
 
       return undefined // Silently swallow errors
     }
+  }
+
+  addIgnoreRule (rule: LanguageToolIgnoredRuleEntry): boolean {
+    const allRules = this._app.config.get().editor.lint.languageTool.ignoredRules
+
+    const existingRule = allRules.find(r => r.id === rule.id)
+    if (existingRule !== undefined) {
+      return false // Don't re-add twice
+    }
+
+    allRules.push(rule)
+    this._app.config.set('editor.lint.languageTool.ignoredRules', JSON.parse(JSON.stringify(allRules)))
+    return true
   }
 }